linter,rules: move type matching code to rules
This code is used in rules-related linter branch,
so it makes sense to move it to the rules package.
